prod builds can run journeycard scripts
