+++ go/client/commands_devel.go
@@ -29,8 +29,7 @@ func getBuildSpecificCommands(cl *libcmdline.CommandLine, g *libkb.GlobalContext
 		newCmdTeamGenerateSeitan(cl, g),
 		newCmdTeamRotateKey(cl, g),
 		newCmdTeamDebug(cl, g),
-		newCmdScript(cl, g),
-		NwCmdContacts(cl, g),
+		NewCmdContacts(cl, g),
 		NewCmdPeopleSearch(cl, g),
 		newCmdTestAirdropReg(cl, g),
 	}
