+++ go/libkb/constants_nonproduction.go
@@ -1,0 +1,8 @@
+// Copyright 2019 Keybase, Inc. All rights reserved. Use of
+// this source code is governed by the included BSD license.
+
+// +build !production
+
+package libkb
+
+const BuildTagProduction = false
